add a party class to inherit from modify dinnerparty and birthday party classto inhaeit from Paety add commaon properties and methods Override Cost and mupda code.
