+++ DinnerParty2.0/DinnerParty.cs
@@ -1,30 +1,26 @@
-﻿using System;
+﻿using Microsoft.SqlServer.Server;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Dinner_Party
+namespace DinnerParty2._0
 {
-    public class DinnerParty
-    {
-        public const int CostOfFoodPerPerson = 25;
+   class DinnerParty:Party
 
-        public int NumberOfPeople { get; set; }
-        public bool FancyDecorations { get; set; }
-        public bool HealthyOption { get; set; }
-        public decimal Cost
+
+    {
+         public bool HealthyOption { get; set; }
+        public override decimal Cost
         {
             get
             {
-                decimal totalCost = 0;
-                totalCost += CalculateCostOfDecorations();
-                totalCost += (CostOfFoodPerPerson + CalculateCostOfBeveragesPerPerson()) * NumberOfPeople;
+                decimal totalCost = base.Cost;
+                totalCost += CalculateCostOfBeveragesPerPerson() * NumberOfPeople;
 
                 if (HealthyOption)
-                {
                     totalCost *= .95m;
-                }
                 return totalCost;
             }
         }
@@ -35,28 +31,13 @@ namespace Dinner_Party
             HealthyOption = healthyOption;
             FancyDecorations = fancyDecorations;
         } 
-
-        private decimal CalculateCostOfDecorations()
-        {
-            if (FancyDecorations)
-            {
-                return (15.00m * NumberOfPeople) + 50m;
-            }
-            else
-            {
-                return (7.50m * NumberOfPeople) + 30m;
-            }
-        }
+         
         private decimal CalculateCostOfBeveragesPerPerson()
         {
             if (HealthyOption)
-            {
                 return 5.00m;
-            }
             else
-            {
                 return 20.00m;
-            }
         }
     }
