+++ DinnerParty2.0/Form1.cs
@@ -1,5 +1,4 @@
-﻿using Dinner_Party;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
